move getStartMtsByTimeframe function to helper
