+++ workers/loc.api/sync/get-risk/helpers/index.js
@@ -2,8 +2,12 @@
 
 const groupByTimeframe = require('./group-by-timeframe')
 const calcGroupedData = require('./calc-grouped-data')
+const getStartMtsByTimeframe = require('./get-start-mts-by-timeframe')
+const getMtsGroupedByTimeframe = require('./get-mts-grouped-by-timeframe')
 
 module.exports = {
   groupByTimeframe,
-  calcGroupedData
+  calcGroupedData,
+  getStartMtsByTimeframe,
+  getMtsGroupedByTimeframe
 }
